Implemented #like and #not_like
As previously stated, this work leans heavily on the stellar work
by @amatsuda and @claudiob - most of this code is, to say the least,
heavily inspired by the code they contributed to rails core.

I've merely extracted it into a separate gem and provided some more
test code (integration tests) as well as a slightly different coding
style.

This has, so far, only been tested with SQLite3 but should be
database-independent, since it builds on the ARel abstraction layer