Limit the number of tests environments required due to combinatorial explosion
As a side effect, we're abandoning guaranteed backwards compatibility. The gem
is only being maintained to work with specific versions of ActiveRecord and no
new features are being made, which means that if you need a version of the gem
to work with an older version of AR, an older version of activerecord-like will
already fulfill your needs.